new-eval-backend/app/negotiate/requirements.txt
azure-functions
azure-functions-signalrservice